refactor: move .database.models import to .database
